Moving session initialisation from mount into initExpressSession
