+++ lib/core/mount.js
@@ -48,14 +48,12 @@
 
 var _ = require('underscore');
 var debug = require('debug')('keystone:core:mount');
-var P = require('bluebird');
 var express = require('express');
 var path = require('path');
 var utils = require('keystone-utils');
 var favicon = require('serve-favicon');
 var morgan = require('morgan');
 var methodOverride = require('method-override');
-var session = require('express-session');
 var multer = require('multer');
 var bodyParser = require('body-parser');
 var cookieParser = require('cookie-parser');
@@ -91,125 +89,8 @@ function mount(mountPath, parentApp, events) {
 		this.set('mongo', dbUrl);
 	}
 
-	// Initialise and validate session options
-
-	if (!this.get('cookie secret')) {
-		console.error('\nKeystoneJS Configuration Error:\n\nPlease provide a `cookie secret` value for session encryption.\n');
-		process.exit(1);
-	}
-
-	var sessionOptions = this.get('session options');
-
-	if (!_.isObject(sessionOptions)) {
-		sessionOptions = {};
-	}
-
-	if (!sessionOptions.key) {
-		sessionOptions.key = 'keystone.sid';
-	}
-
-	_.defaults(sessionOptions, {
-		resave: false,
-		secret: this.get('cookie secret'),
-		saveUninitialized: false
-	});
-
-	sessionOptions.cookieParser = cookieParser(this.get('cookie secret'));
-
-	var sessionStore = this.get('session store');
-
-	var sessionStorePromise;
-
-	if (typeof sessionStore === 'function') {
-		sessionOptions.store = sessionStore(session);
-	} else if (sessionStore) {
-
-		var sessionStoreOptions = this.get('session store options') || {};
-
-		// Perform any session store specific configuration or exit on an unsupported session store
-
-		if (sessionStore === 'mongo') {
-			sessionStore = 'connect-mongo';
-		} else if (sessionStore === 'redis') {
-			sessionStore = 'connect-redis';
-		}
-
-		switch (sessionStore) {
-			case 'connect-mongo':
-				debug('using mongo for session store');
-				_.defaults(sessionStoreOptions, {
-					collection: 'app_sessions',
-					url: this.get('mongo')
-				});
-				break;
-
-			case 'connect-mongostore':
-				debug('using connect-mongostore for session store');
-				_.defaults(sessionStoreOptions, {
-					collection: 'app_sessions'
-				});
-				if (!sessionStoreOptions.db) {
-					console.error(
-						'\nERROR: connect-mongostore requires `session store options` to be set.' +
-						'\n' +
-						'\nSee http://keystonejs.com/docs/configuration#options-database for details.' +
-					'\n');
-					process.exit(1);
-				}
-				break;
-
-			case 'connect-redis':
-				debug('using redis for session store');
-				break;
-
-			default:
-				console.error(
-					'\nERROR: unsupported session store ' + sessionStore + '.' +
-					'\n' +
-					'\nSee http://keystonejs.com/docs/configuration#options-database for details.' +
-				'\n');
-				process.exit(1);
-				break;
-		}
-
-		// Initialize the session store
-		try {
-			var SessionStore = require(sessionStore)(session);
-			sessionStorePromise = new P(function(resolve) {
-				sessionOptions.store = new SessionStore(sessionStoreOptions, resolve);
-				sessionOptions.store.on('connect', resolve);
-				sessionOptions.store.on('connected', resolve);
-				sessionOptions.store.on('disconnect', function() {
-					console.error(
-						'\nThere was an error connecting to the ' + sessionStore + ' session store.' +
-					'\n');
-					process.exit(1);
-				});
-			});
-		} catch(e) {
-			if (e.code === 'MODULE_NOT_FOUND') {
-				console.error(
-					'\n' + e.toString() +
-					'\nTo use ' + this.get('session store') + ' as a `session store` option, run:' +
-					'\nnpm install ' + sessionStore + ' --save' +
-				'\n');
-				process.exit(1);
-			} else {
-				throw e;
-			}
-		}
-	}
-
-	// expose initialised session and options
-	this.set('session options', sessionOptions);
-	this.set('express session', session(sessionOptions));
-
-
-	// wrangle arguments
-	if (arguments.length === 1) {
-		events = arguments[0];
-		mountPath = null;
-	}
+	// Initialise the session store
+	this.initExpressSession();
 
 	if ('function' === typeof events) {
 		events = { onMount: events };
@@ -390,7 +271,7 @@ function mount(mountPath, parentApp, events) {
 		keystone.callHook('pre:session', req, res, next);
 	});
 
-	app.use(this.get('express session'));
+	app.use(this.expressSession);
 	app.use(multer({
 		includeEmptyFields: true
 	}));
@@ -634,8 +515,8 @@ function mount(mountPath, parentApp, events) {
 			}
 		};
 
-		if (sessionStorePromise) {
-			sessionStorePromise.then(connected);
+		if (this.sessionStorePromise) {
+			this.sessionStorePromise.then(connected);
 		} else {
 			connected();
 		}
